% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate-distances.R
\name{calculate_distance}
\alias{calculate_distance}
\title{Calculate pairwise distances between sequences using various methods}
\usage{
calculate_distance(
  x,
  method,
  subtype = NULL,
  tree = NULL,
  map = NULL,
  substitution_matrix = NULL,
  ...
)
}
\arguments{
\item{x}{A character vector of sequences (typically amino acid or nucleotide
sequences). Currently, you can pass NULL if you want to use the
`"cartographic"` or `"cophenetic"` methods without specifying a sequence
vector.}

\item{method}{A character string specifying the distance method to use.
Available options include:
\itemize{
  \item `"hamming"`: Computes Hamming distance.
  \item `"dl"`: Computes Damerau-Levenshtein distance.
  \item `"p-epitope"`: Computes p-epitope distances for a specified subtype, by default the dominant p-peitope distance is returned but you can access all arguments of \code{\link{dist_pepi}}.
  \item `"p-all-epitope"`: Computes p-all-epitope distances for a subtype.
  \item `"cophenetic"` or `"tree"`: Computes distances based on a phylogenetic tree.
  \item `"cartographic"` or `"cartography"`: Computes cartographic distances from a Racmacs map.
  \item `"temporal"`: Computes temporal distances based on sequence names.
  \item `"grantham"`: Computes Grantham's distance based on a substitution matrix.
  \item `"substitution"`: Computes general weighted Hamming distances based on different substitution metrics which you can specify with the `substitution_matrix` argument.
}}

\item{subtype}{(Optional) A character string indicating the subtype (e.g.,
`"H1"`, `"B"`) to use when computing p-epitope distances. Required for
methods `"p-epitope"` and `"p-all-epitope"`.}

\item{tree}{(Optional) A `phylo` object used when `method = "cophenetic"` or
`"tree"`.}

\item{map}{(Optional) A Racmacs map object used when `method =
"cartographic"` or `"cartography"`.}

\item{...}{Additional arguments passed to the corresponding distance
functions.}
}
\value{
A numeric matrix of pairwise distances between the input sequences.
}
\description{
This function provides a unified interface for computing pairwise distances
between sequences using a variety of distance metrics. It dispatches to
appropriate internal functions based on the specified method.
}
\details{
This function acts as a dispatcher, validating the `method` input
  and routing the data to the correct underlying distance function. Required
  arguments for specific methods (e.g., `subtype`, `tree`, `map`) must be
  supplied when applicable or an error will be thrown.
}
\examples{
\dontrun{calculate_distance(my_seq, "hamming")}
\dontrun{calculate_distance(my_seq, "p-epitope", subtype = "A(H1N1)")}
\dontrun{calculate_distance(my_seq, "cartographic", map = my_ac_map)}

}
\seealso{
\code{\link{dist_string}} for `"hamming"` and `"dl"` distances,
  \code{\link{dist_pepi}} for `"p-epitope"` and `"p-all-epitope"` distances,
  \code{\link{tree_to_distances}} for `"cophenetic"` or `"tree"` distances,
  \code{\link{racmaps_map_to_distances}} for `"cartographic"` or `"cartography"` distances,
  \code{\link{dist_temporal}} for `"temporal"` distances.
  \code{\link{dist_substitution}} for `"substitution"` and `"grantham"` distances.
}
