% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/string-based.R
\name{dist_string}
\alias{dist_string}
\title{Compute Pairwise Normalized String Distances Between Aligned Sequences}
\usage{
dist_string(seqs, ambiguous_residues = "xX?", ...)
}
\arguments{
\item{seqs}{A character vector or list of aligned sequences.}

\item{ambiguous_residues}{A character vector of ambiguous residues to remove
from each sequence before comparison. Defaults to \code{c("x", "X", "?")}.}

\item{...}{Additional arguments passed to \code{stringdist::stringdist()}
(e.g., \code{method = "hamming"}).}
}
\value{
A symmetric numeric matrix of pairwise normalized string distances.
  Each distance is normalized by the length of the cleaned first sequence in
  the pair.
}
\description{
This function computes a pairwise distance matrix between aligned sequences
using a string distance metric from the `stringdist` package. Ambiguous
residues (e.g., "X", "x", "?") are removed before distance computation.
By default, the hamming distance is compouted, but the
\code{stringdist::stringdist()} documentation lists other possible methods
and arguments.
}
\details{
Only the lower triangle is explicitly computed, and the upper
  triangle is filled in by symmetry. This function assumes that all sequences
  are of equal length and aligned.

  All distances are normalized by dividing by the aligned sequence length.
}
\examples{
seqs <- c(
  "A/H1N1/South Carolina/1/1918" = "mktiialsyifclvlgqdfpgndnstat",
  "A/H3N2/Darwin/9/2021" = "mktiialsnilclvfaqkipgndnstat",
  "B/Sichuan/379/1999" = "drictgitssnsphvvktatqgevnvtg"
)
dist_string(seqs, method = "hamming")

}
