% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{tidy_dist_mat}
\alias{tidy_dist_mat}
\title{Convert a distance matrix to tidy format}
\usage{
tidy_dist_mat(
  d,
  rows_to = "Strain1",
  cols_to = "Strain2",
  unique_pairs = FALSE
)
}
\arguments{
\item{d}{A square numeric matrix (or coercible to one) with row and column
names.}

\item{rows_to}{A string (length 1) naming the new column for row labels.
Default is `"Strain1"`.}

\item{cols_to}{A string (length 1) naming the new column for column labels.
Default is `"Strain2"`.}

\item{unique_pairs}{Logical (length 1). If `TRUE`, return only the unique
row-column combinations from the lower triangle and diagonal of the matrix.
Defaults to `FALSE`.}
}
\value{
A tibble in long format with columns:
\describe{
  \item{[rows_to]}{Factor, ordered by the row order of `d`}
  \item{[cols_to]}{Factor, ordered and reversed from column order of `d`}
  \item{d}{Distance value between row and column entries}
}
}
\description{
Transforms a square numeric distance matrix into a long-format tibble with
customizable names for the row and column variables. Optionally, returns only
unique row-column combinations (i.e., lower triangle).
}
\examples{
mat <- matrix(1:9, nrow = 3)
rownames(mat) <- colnames(mat) <- c("A", "B", "C")
tidy_dist_mat(mat)
tidy_dist_mat(mat, unique_pairs = TRUE)

}
