% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model-based.R
\name{tree_to_distances.default}
\alias{tree_to_distances.default}
\title{Default method for `tree_to_distances`}
\usage{
\method{tree_to_distances}{default}(x, ...)
}
\arguments{
\item{x}{An object of unsupported class.}

\item{...}{Not used.}
}
\value{
No return value; prints a message to the console.
}
\description{
This method is called when `tree_to_distances()` is used on an unsupported
object class. It prints an informative message listing available methods.
}
