% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model-based.R
\name{tree_to_distances.phylo}
\alias{tree_to_distances.phylo}
\title{Extract distances from a phylo object}
\usage{
\method{tree_to_distances}{phylo}(x, ...)
}
\arguments{
\item{x}{A phylogenetic tree of class `phylo` (from the `ape` package).}

\item{...}{Not used.}
}
\value{
An object of class `dist`, with the distances sorted by tip labels.
}
\description{
Computes the pairwise cophenetic distance matrix from a phylogenetic tree of
class `phylo`. The output is sorted by tip labels.
}
